requirements.txt
streamlit>=1.37
pandas>=2.2
numpy>=1.26
matplotlib>=3.8
python-dotenv>=1.0
openai>=1.14
